Ports `Conv` and `ConvTranspose` to TRT dialect, makes Modules update shapes
- Ports the `Conv` and `ConvTranspose` modules to use the TRT dialect.

- Updates `Module` to update the shapes of parameter trace tensors when it
    computes them to check compatibility. This makes the trace more informative
    and is also required for functional correctness in some operations.
