+++ tripy/nvtripy/flat_ir/ops/__init__.py
@@ -23,7 +23,6 @@ from nvtripy.flat_ir.ops.compare import CompareOp
 from nvtripy.flat_ir.ops.concatenate import ConcatenateOp
 from nvtripy.flat_ir.ops.constant import ConstantOp
 from nvtripy.flat_ir.ops.convert import ConvertOp
-from nvtripy.flat_ir.ops.convolution import ConvolutionOp
 from nvtripy.flat_ir.ops.divide import DivideOp
 from nvtripy.flat_ir.ops.dot import DotOp
 from nvtripy.flat_ir.ops.floor import FloorOp
